refactor: 메서드 통합 및 클래스 분리
로또 생성 중 정렬 부분을 로또 도메인으로 통합, 입력값을 받아 변환하는 부분을 새로운 클래스로 분리

Ref: #2
